Add hook to augment `Platform` objects for Artifact selection
This, combined with the recent overhaul of BinaryPlatforms to give
extended tags to `Platform` objects, enables packages to run arbitrary
(ideally, small) pieces of code to better inform artifact selection.

`Platform`-specific Artifacts can now be tagged with arbitrary metadata,
examples of which are `libc` version, processor microarchitecture, or
CUDA version.  Because we do not want to have to teach Julia how to
autodetect its compatibility with all possible tags, we instead add the
ability for a package to perform the autodetection, and run that code
after the package is installed, but before its artifacts are downloaded.

Cooperating packages will create a `.pkg/platform_augmentation_hook.jl`
file within their package, and that file will return a function as its
final value.  This code will be evaluated within an anonymous module,
and within the Package's own environment, allowing the hook to import
functionality if need be (this should be avoided as much as possible, so
as to reduce the possibility of conflicting versions of dependencies
causing an issue at install time).